+++ app/helpers/application_helper.rb
@@ -35,7 +35,7 @@ module ApplicationHelper
   end
 
   def chapter_students_number(chapter)
-    chapter.groups.reduce(0) { |acc, elem| acc + elem.students.length }
+    chapter.groups.reduce(0) { |acc, elem| acc + elem.students.exclude_deleted.length }
   end
 
   def user_image(user, size = 40)